prelude and reorganization
