+++ src/cakes/mod.rs
@@ -1,0 +1,3 @@
+pub use search::Search;
+
+mod search;
