+++ src/manifold.rs
@@ -1,10 +1,7 @@
 use std::sync::Arc;
 
-use crate::cluster::Cluster;
 use crate::criteria::ClusterCriterion;
-use crate::dataset::Dataset;
-use crate::graph::Graph;
-use crate::metric::Number;
+use crate::prelude::*;
 
 pub struct Manifold<T: Number, U: Number> {
     pub dataset: Arc<dyn Dataset<T, U>>,
