+++ src/prelude.rs
@@ -2,8 +2,12 @@ use std::sync::Arc;
 
 use dashmap::{DashMap, DashSet};
 
-use crate::cluster::Cluster;
-use crate::graph::Edge;
+pub use crate::cluster::Cluster;
+pub use crate::dataset::Dataset;
+pub use crate::graph::Edge;
+pub use crate::graph::Graph;
+pub use crate::manifold::Manifold;
+pub use crate::metric::{metric_new, Metric, Number};
 
 pub type Index = usize;
 pub type Indices = Vec<Index>;
